LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/window-open.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py


Harness Error (TIMEOUT), message = null

PASS Same-origin window, forced
TIMEOUT Same-site window, forced Test timed out
FAIL Cross-site window, forced The string did not match the expected pattern.
PASS Same-origin window, forced, reloaded
TIMEOUT Same-site window, forced, reloaded Test timed out
FAIL Cross-site window, forced, reloaded The string did not match the expected pattern.
PASS Same-origin window, user-activated
TIMEOUT Same-site window, user-activated Test timed out
FAIL Cross-site window, user-activated The string did not match the expected pattern.
PASS Same-origin window, forced: sec-fetch-dest
PASS Same-origin window, forced: sec-fetch-mode
PASS Same-origin window, forced: sec-fetch-site
PASS Same-origin window, forced: sec-fetch-user
PASS Same-origin window, user-activated: sec-fetch-dest
PASS Same-origin window, user-activated: sec-fetch-mode
PASS Same-origin window, user-activated: sec-fetch-site
FAIL Same-origin window, user-activated: sec-fetch-user assert_equals: expected "?1" but got ""
PASS Same-origin window, forced, reloaded: sec-fetch-dest
PASS Same-origin window, forced, reloaded: sec-fetch-mode
PASS Same-origin window, forced, reloaded: sec-fetch-site
PASS Same-origin window, forced, reloaded: sec-fetch-user
